Move getRenamedMGU to FindMGU
